HBX-1592: Move class 'org.hibernate.tool.util.TableNameQualifier' to 'org.hibernate.tool.internal.util.TableNameQualifier'
