+++ main/src/main/java/org/hibernate/cfg/reveng/IndexProcessor.java
@@ -15,7 +15,7 @@ import org.hibernate.mapping.Table;
 import org.hibernate.mapping.UniqueKey;
 import org.hibernate.tool.api.dialect.MetaDataDialect;
 import org.hibernate.tool.internal.reveng.JdbcBinderException;
-import org.hibernate.tool.util.TableNameQualifier;
+import org.hibernate.tool.internal.util.TableNameQualifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
