+++ main/src/main/java/org/hibernate/tool/internal/dialect/HSQLMetaDataDialect.java
@@ -8,7 +8,7 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 
-import org.hibernate.tool.util.TableNameQualifier;
+import org.hibernate.tool.internal.util.TableNameQualifier;
 
 /**
  * @author Dmitry Geraskov
