+++ main/src/main/java/org/hibernate/tool/internal/dialect/JDBCMetaDataDialect.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 
-import org.hibernate.tool.util.TableNameQualifier;
+import org.hibernate.tool.internal.util.TableNameQualifier;
 
 /**
  * MetaData dialect that uses standard JDBC for reading metadata.
