+++ main/src/main/java/org/hibernate/tool/internal/util/TableNameQualifier.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.util;
+package org.hibernate.tool.internal.util;
 
 public class TableNameQualifier {
